% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_external_resources.R
\name{add_external_resources}
\alias{add_external_resources}
\title{Add External Resources to the Application
This function is internally used to add external resources inside the Shiny
application.}
\usage{
add_external_resources()
}
\description{
Add External Resources to the Application
This function is internally used to add external resources inside the Shiny
application.
}
