% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_description.R
\name{convert_md_to_html}
\alias{convert_md_to_html}
\title{Convert Strategy Text from Markdown to HTML}
\usage{
convert_md_to_html(text)
}
\arguments{
\item{text}{Character. The Markdown text description for a strategy, as read
by \link{fetch_strategy_description}.}
}
\value{
HTML/character.
}
\description{
Convert Strategy Text from Markdown to HTML
}
