% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpma-explorer-app.R
\docType{package}
\name{cpma.explorer-package}
\alias{cpma.explorer}
\alias{cpma.explorer-package}
\title{cpma.explorer: Explore CPMA Data}
\description{
A Shiny-app-in-a-package to explore data related to categories of potentially mitigaable activity (CPMAs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/The-Strategy-Unit/cpma-explorer}
  \item Report bugs at \url{https://github.com/The-Strategy-Unit/cpma-explorer/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{matt.dray@nhs.net}

}
\keyword{internal}
