% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{download_all_data}
\alias{download_all_data}
\title{Download Inputs Datasets for Specific Geography}
\usage{
download_all_data(
  data_version = Sys.getenv("DATA_VERSION", "dev"),
  redownload = FALSE
)
}
\description{
Download Inputs Datasets for Specific Geography
}
