% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{download_geo_data}
\alias{download_geo_data}
\title{Download Inputs Datasets for Specific Geography}
\usage{
download_geo_data(geography)
}
\arguments{
\item{geography}{Character. The geography level for which the user wants to
select a provider. Either "nhp" or "la".}
}
\value{
A list. The filenames of the downloaded datasets.
}
\description{
Download Inputs Datasets for Specific Geography
}
