% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{enstring_fyear}
\alias{enstring_fyear}
\title{Convert a Financial Year to Character}
\usage{
enstring_fyear(year)
}
\arguments{
\item{year}{Integer. Expressed in the form \code{202526}. To be converted to the
form \code{"2025/26"}.}
}
\description{
Convert a Financial Year to Character
}
