% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{entable_diagnoses}
\alias{entable_diagnoses}
\title{Create 'gt' Summary Table of Diagnoses}
\usage{
entable_diagnoses(diagnoses_prepared)
}
\arguments{
\item{diagnoses_prepared}{A data.frame.}
}
\value{
A 'gt' table.
}
\description{
Create 'gt' Summary Table of Diagnoses
}
