% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table.R
\name{entable_encounters}
\alias{entable_encounters}
\title{Create 'gt' Summary Table of Procedures or Diagnoses}
\usage{
entable_encounters(encounters_prepared)
}
\arguments{
\item{encounters_prepared}{A data.frame. Procedures or diagnoses
('encounters') data read from Azure and processed with
\link{prepare_procedures_data} or \link{prepare_diagnoses_data}.}
}
\value{
A 'gt' table.
}
\description{
Create 'gt' Summary Table of Procedures or Diagnoses
}
