% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{entable_procedures}
\alias{entable_procedures}
\title{Create 'gt' Summary Table of Procedures}
\usage{
entable_procedures(procedures_prepared)
}
\arguments{
\item{procedures_prepared}{A data.frame.}
}
\value{
A 'gt' table.
}
\description{
Create 'gt' Summary Table of Procedures
}
