% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_show_strategy_text.R
\name{fetch_strategy_text}
\alias{fetch_strategy_text}
\title{Fetch Strategy Text from NHP Inputs}
\usage{
fetch_strategy_text(strategy_stub, filename)
}
\arguments{
\item{strategy_stub}{Character. Strategy variable name stub, e.g.
\code{"alcohol_partially_attributable"}.}

\item{filename}{Character. The filename to save the fetched strategy text to.}
}
\description{
Fetch Strategy Text from NHP Inputs
}
