% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{generate_rates_funnel_data}
\alias{generate_rates_funnel_data}
\title{Generate Rates Baseline Data}
\usage{
generate_rates_funnel_data(
  selected_geography,
  selected_provider,
  selected_strategy,
  selected_year,
  providers_lookup,
  peers_lookup
)
}
\arguments{
\item{selected_geography}{Character. Selected geography, either \code{"nhp"} or \code{"la"}.}

\item{selected_provider}{Character. Provider code, e.g. \code{"RCF"}.}

\item{selected_strategy}{Character. Strategy variable name, e.g. \code{"alcohol_partially_attributable_acute"}.}

\item{selected_year}{Integer. Selected year in the form \code{202324}.}

\item{providers_lookup}{Dataframe. A lookup from a provider code to its name.}

\item{peers_lookup}{Character. A vector of peer provider codes.}
}
\value{
A data.frame.
}
\description{
Generate Rates Baseline Data
}
