% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{generate_rates_trend_data}
\alias{generate_rates_trend_data}
\title{Generate Rates Trend Data}
\usage{
generate_rates_trend_data(
  selected_geography,
  selected_provider,
  selected_strategy
)
}
\arguments{
\item{selected_geography}{Character. Selected geography, either \code{"nhp"} or \code{"la"}.}

\item{selected_provider}{Character. Provider code, e.g. \code{"RCF"}.}

\item{selected_strategy}{Character. Strategy variable name, e.g. \code{"alcohol_partially_attributable_acute"}.}
}
\value{
A data.frame.
}
\description{
Generate Rates Trend Data
}
