% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_all_geo_data}
\alias{get_all_geo_data}
\title{Read Inputs Datasets for All Geographies}
\usage{
get_all_geo_data(geography, fyear)
}
\arguments{
\item{geography}{Character. The geography level for which the user wants to
select a provider. Either "nhp" or "la".}

\item{fyear}{Numeric. The financial year to filter the data to.}
}
\value{
A list. One element for each dataframes of data.
}
\description{
Read Inputs Datasets for All Geographies
}
