% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_arrow_dataset}
\alias{get_arrow_dataset}
\title{Get an Arrow Dataset}
\usage{
get_arrow_dataset(geography, dataset_name)
}
\arguments{
\item{geography}{Character. Either "nhp" or "la", used to determine the file
path.}

\item{dataset_name}{Character. The name of the dataset, e.g. "rates".}
}
\value{
An Arrow dataset object.
}
\description{
This function loads a parquet file as an Arrow dataset and caches it in
memory for future use.
}
