% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_rates.R
\name{get_peers_lookup}
\alias{get_peers_lookup}
\title{Get Peers Lookup}
\usage{
get_peers_lookup(selected_geography)
}
\arguments{
\item{selected_geography}{Character. The selected geography, e.g. "nhp" or "la".}
}
\value{
A data.frame with provider codes and their corresponding peers.
}
\description{
Reads the appropriate peers lookup file based on the selected geography.
}
