% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_rates.R
\name{get_rates_data}
\alias{get_rates_data}
\title{Get Rates Data}
\usage{
get_rates_data(df, strategy)
}
\arguments{
\item{df}{the rates data from inputs_data}

\item{strategy}{Character. Strategy variable name to filter rows to}
}
\value{
Dataframe with national rates in a separate column
}
\description{
This function prepares the data for the rates_data() reactive. It moves the
national row to a separate column.
}
