% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_rates.R
\name{get_rates_trend_data}
\alias{get_rates_trend_data}
\title{Get Rates Trend Data}
\usage{
get_rates_trend_data(df, provider)
}
\arguments{
\item{df}{A data.frame. Prepared by get_rates_data}

\item{provider}{Character. Provider code, e.g. \code{"RCF"}.}
}
\value{
Dataframe filtered to the specific provider
}
\description{
Extracts the trend for a specific provider from the rates data
(see \link{get_rates_data}).
}
