% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_rates.R
\name{get_rates_y_axis_limits}
\alias{get_rates_y_axis_limits}
\title{Get Rates Y Axis Limits}
\usage{
get_rates_y_axis_limits(
  rates_trend_data,
  rates_funnel_data,
  rates_funnel_calculations
)
}
\arguments{
\item{rates_trend_data}{A data.frame. Prepared by get_rates_trend_data.}

\item{rates_funnel_data}{A data.frame. Prepared by generate_rates_funnel_data.}

\item{rates_funnel_calculations}{A data.frame. Prepared by uprime_calculations.}
}
\value{
A numeric vector of length 2, representing the lower and upper limits of the y-axis.
}
\description{
Determines the y-axis limits for the rates trend plot based on the trend data and funnel data.
}
