% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{isolate_provider_peers}
\alias{isolate_provider_peers}
\title{Find the Peers for a Given Provider}
\usage{
isolate_provider_peers(provider, peers)
}
\arguments{
\item{provider}{Character. Provider code, e.g. \code{"RCF"}.}

\item{peers}{A data.frame. A row per provider-peer pair.}
}
\value{
Character vector of peers for given \code{provider}.
}
\description{
Find the Peers for a Given Provider
}
