% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{make_strategy_group_lookup}
\alias{make_strategy_group_lookup}
\title{Create a Simple Lookup of Strategies to Strategy Groups}
\usage{
make_strategy_group_lookup(strategies_config)
}
\arguments{
\item{strategies_config}{List. Configuration for strategies from the
\code{"mitigators_config"} element of \code{golem-config.yml}, read in with
\link{get_golem_config}.}
}
\description{
Create a Simple Lookup of Strategies to Strategy Groups
}
