% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{md_file_to_html}
\alias{md_file_to_html}
\title{Read an Markdown File and Convert to HTML}
\usage{
md_file_to_html(...)
}
\arguments{
\item{...}{Character vectors. Construct a path to a Markdown file (like
\link{file.path}).}
}
\value{
A data.frame.
}
\description{
Read an Markdown File and Convert to HTML
}
