% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{md_string_to_html}
\alias{md_string_to_html}
\title{Convert a Markdown String to HTML}
\usage{
md_string_to_html(text)
}
\arguments{
\item{text}{Character string. Markdown text to convert to HTML.}
}
\value{
A shiny HTML object.
}
\description{
Convert a Markdown String to HTML
}
