% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_funnel}
\alias{plot_funnel}
\title{Plot Funnel}
\usage{
plot_funnel(rates_funnel_data, plot_range, x_axis_title)
}
\arguments{
\item{rates_funnel_data}{A data.frame.}

\item{plot_range}{Character. A vector of length two giving the y-axis limits.}

\item{x_axis_title}{Character.}
}
\value{
A ggplot2 object.
}
\description{
Plot Funnel
}
