% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_nee}
\alias{plot_nee}
\title{Plot National Elicitation Exercise (NEE) Interval}
\usage{
plot_nee(nee_data)
}
\arguments{
\item{nee_data}{A data.frame. Mean, p10 and p90 predictions for each of the
strategies that were part of the National Elicitation Exercise (NEE),
filtered for the user's selected strategy.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot National Elicitation Exercise (NEE) Interval
}
