% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_rates_box}
\alias{plot_rates_box}
\title{Plot Rates Boxplot with Peers}
\usage{
plot_rates_box(rates_box_data, y_axis_limits, base_size = 16)
}
\arguments{
\item{rates_box_data}{A data.frame. Rates data}

\item{y_axis_limits}{Numeric vector. Min and max values for the y axis.}

\item{base_size}{Numeric scalar. For scaling plot-element sizes.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Rates Boxplot with Peers
}
