% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_rates.R
\name{plot_rates_funnel}
\alias{plot_rates_funnel}
\title{Plot Rates Funnel with Peers}
\usage{
plot_rates_funnel(rates_funnel_data, plot_range, x_axis_title)
}
\arguments{
\item{rates_funnel_data}{A data.frame.}

\item{plot_range}{Character. A vector of length two giving the y-axis limits.}

\item{x_axis_title}{Character.}
}
\value{
A ggplot2 object.
}
\description{
Plot Rates Funnel with Peers
}
