% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{prepare_age_sex_data}
\alias{prepare_age_sex_data}
\title{Prepare Age-Sex Data}
\usage{
prepare_age_sex_data(geography, provider, strategy, year)
}
\arguments{
\item{geography}{Character. Either "nhp" or "la", used to determine the file path.}

\item{provider}{Character. Provider identifier.}

\item{strategy}{Character. Strategy identifier.}

\item{year}{Numeric. Financial year.}
}
\value{
A data.frame.
}
\description{
Prepare Age-Sex Data
}
