% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{prepare_diagnoses_data}
\alias{prepare_diagnoses_data}
\title{Prepare Data for Diagnoses Table}
\usage{
prepare_diagnoses_data(
  diagnoses_lookup,
  geography,
  provider,
  strategy,
  selected_year
)
}
\arguments{
\item{diagnoses_lookup}{A data.frame. Type, code and description for
diagnoses}

\item{geography}{Character. Geography type, either \code{"nhp"} or \code{"la"}.}

\item{provider}{Character. Provider code, e.g. \code{"RCF"}.}

\item{strategy}{Character. Strategy variable name, e.g.
\code{"alcohol_partially_attributable_acute"}.}

\item{selected_year}{Integer. Baseline year in the form \code{202324}.}
}
\value{
A data.frame.
}
\description{
Prepare Data for Diagnoses Table
}
