% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tables.R
\name{prepare_procedures_data}
\alias{prepare_procedures_data}
\title{Prepare Data for Procedures Table}
\usage{
prepare_procedures_data(
  procedures_data,
  procedures_lookup,
  provider,
  strategy,
  baseline_year
)
}
\arguments{
\item{procedures_data}{A data.frame. Procedure data read in from Azure.
Annual procedure counts by provider and strategy.}

\item{procedures_lookup}{A data.frame. Type, code and description for
procedures.}

\item{provider}{Character. Provider code, e.g. \code{"RCF"}.}

\item{strategy}{Character. Strategy variable name, e.g.
\code{"alcohol_partially_attributable_acute"}.}

\item{baseline_year}{Integer. Baseline year in the form \code{202324}.}
}
\value{
A data.frame.
}
\description{
Prepare Data for Procedures Table
}
