% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_show_strategy_text.R
\name{read_strategy_text}
\alias{read_strategy_text}
\title{Read Strategy Text from File}
\usage{
read_strategy_text(strategy_stub)
}
\arguments{
\item{strategy_stub}{Character. Strategy variable name stub, e.g.
\code{"alcohol_partially_attributable"}.}
}
\value{
A character vector containing the lines of the strategy text.
}
\description{
Reads the strategy text from a local file. If the file does not exist, it
fetches the strategy text from the remote repository and saves it to the file
before reading it.
}
\details{
Markdown files containing strategy descriptions are read from
\href{https://github.com/The-Strategy-Unit/nhp_inputs/}{NHP Inputs}.
}
