% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{theme_base}
\alias{theme_base}
\title{A 'ggplot2' Theme for All Plots}
\usage{
theme_base(base_size = 16)
}
\arguments{
\item{base_size}{Numeric. Size of base text from which other sizes can be
calculated.}
}
\value{
A 'ggplot2' theme.
}
\description{
A 'ggplot2' Theme for All Plots
}
