% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpma-explorer-app.R
\docType{package}
\name{tpma.explorer-package}
\alias{tpma.explorer}
\alias{tpma.explorer-package}
\title{tpma.explorer: Explore Opportunities to Reduce Hospital Care}
\description{
A Shiny-app-in-a-package to explore data related to Types of Potentially-Mitigatable Activity (TPMAs) with the goal of finding opportunities to reduce hospital care.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/The-Strategy-Unit/tpma-explorer}
  \item Report bugs at \url{https://github.com/The-Strategy-Unit/tpma-explorer/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{matt.dray@nhs.net}

Authors:
\itemize{
  \item Tom Jemmett \email{thomas.jemmett@nhs.net}
}

Other contributors:
\itemize{
  \item Claire Welsh \email{claire.welsh8@nhs.net} [contributor]
}

}
\keyword{internal}
