% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{uprime_calculations}
\alias{uprime_calculations}
\title{Generate Function to Calculate U-Prime values}
\usage{
uprime_calculations(df)
}
\arguments{
\item{rates_baseline_data}{A data.frame. Rates data read in from Azure and
processed \link{generate_rates_baseline_data}.}
}
\value{
A list containing items to produce a U-Prime funnel chart.
}
\description{
Generate Function to Calculate U-Prime values
}
