% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{download_all_data}
\alias{download_all_data}
\title{Download Inputs Datasets for Specific Geography}
\usage{
download_all_data(
  data_version = Sys.getenv("DATA_VERSION", "dev"),
  redownload = FALSE
)
}
\arguments{
\item{data_version}{Character. The version of the data to download. By
default, uses the value of the "DATA_VERSION" environment variable, or
"dev" if that variable is not set.}

\item{redownload}{Logical. Whether to redownload the data if it already
exists. By default, FALSE.}
}
\description{
Downloads all datasets for both "provider" and "lad23cd" geographies. See
\code{download_geo_data()} for more details.
}
