% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_show_strategy_text.R
\name{fetch_strategy_text}
\alias{fetch_strategy_text}
\title{Fetch Strategy Text from NHP Inputs}
\usage{
fetch_strategy_text(strategy, descriptions_lookup)
}
\arguments{
\item{strategy}{Character. The variable name for a strategy, e.g.
\code{"discharged_no_treatment_adult_ambulance"}.}

\item{descriptions_lookup}{Character. The names of the available Markdown
description files.}
}
\value{
Character.
}
\description{
Fetch Strategy Text from NHP Inputs
}
\details{
Markdown files containing strategy descriptions are read from
\href{https://github.com/The-Strategy-Unit/nhp_inputs/}{NHP Inputs}.
}
