% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_show_strategy_text.R
\name{fetch_strategy_text}
\alias{fetch_strategy_text}
\title{Fetch Strategy Text from NHP Inputs}
\usage{
fetch_strategy_text(strategy_stub)
}
\arguments{
\item{strategy_stub}{Character. Strategy variable name stub, e.g.
\code{"alcohol_partially_attributable"}.}
}
\value{
Character.
}
\description{
Fetch Strategy Text from NHP Inputs
}
\details{
Markdown files containing strategy descriptions are read from
\href{https:","","github.com","The-Strategy-Unit","nhp_inputs","}{NHP Inputs}.
}
