% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{generate_rates_baseline_data}
\alias{generate_rates_baseline_data}
\title{Generate Rates Baseline Data}
\usage{
generate_rates_baseline_data(rates, provider, peers_lookup, selected_year)
}
\arguments{
\item{rates}{A data.frame. Rates data read from Azure.}

\item{provider}{Character. Provider code, e.g. \code{"RCF"}.}

\item{peers_lookup}{Dataframe. A lookup from a provider to it's peers.}

\item{selected_year}{Integer. Baseline year in the form \code{202324}.}
}
\value{
A data.frame.
}
\description{
Generate Rates Baseline Data
}
