% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{generate_rates_baseline_data}
\alias{generate_rates_baseline_data}
\title{Generate Rates Baseline Data}
\usage{
generate_rates_baseline_data(rates, provider, peers, strategy, selected_year)
}
\arguments{
\item{rates}{A data.frame. Rates data read from Azure.}

\item{provider}{Character. Provider code, e.g. \code{"RCF"}.}

\item{peers}{Character. A vector of peers for given \code{provider}.}

\item{strategy}{Character. Strategy variable name, e.g.
\code{"alcohol_partially_attributable_acute"}.}

\item{selected_year}{Integer. Baseline year in the form \code{202324}.}
}
\value{
A data.frame.
}
\description{
Generate Rates Baseline Data
}
