% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_rates.R
\name{generate_rates_baseline_data}
\alias{generate_rates_baseline_data}
\title{Generate Rates Baseline Data}
\usage{
generate_rates_baseline_data(rates, provider, peers, strategy, start_year)
}
\arguments{
\item{rates}{A data.frame.}

\item{provider}{Character. Provider code, e.g. \code{"RCX"}.}

\item{peers}{Character. A vector of peers for given \code{provider}.}

\item{strategy}{Character.  TPMA variable name, e.g.
\code{"alcohol_partially_attributable_acute"}.}

\item{start_year}{Integer. Baseline year in the form \code{202324}.}
}
\value{
A data.frame.
}
\description{
Generate Rates Baseline Data
}
