% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{generate_rates_funnel_data}
\alias{generate_rates_funnel_data}
\title{Generate Data for the Funnel Plot}
\usage{
generate_rates_funnel_data(rates_baseline_data)
}
\arguments{
\item{rates_baseline_data}{A data.frame. Output created by
\link{generate_rates_baseline_data}.}
}
\value{
A data.frame.
}
\description{
Generate Data for the Funnel Plot
}
