% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_all_geo_data}
\alias{get_all_geo_data}
\title{Read Inputs Datasets for All Geographies}
\usage{
get_all_geo_data(geographies, data_types)
}
\arguments{
\item{geographies}{Character. The geography level for which the user wants to
select a provider.}

\item{data_types}{Character. A vector of filenames (without filetypes) for
each of the inputs data files that you want to read from
the \code{inputs_container}.}
}
\value{
A list. One element for each of the \code{geographies}, with subelements
that are dataframes of each one of the \code{data_types}.
}
\description{
Read Inputs Datasets for All Geographies
}
