% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure.R
\name{get_container}
\alias{get_container}
\title{Get Azure Container}
\usage{
get_container(
  ep_uri = Sys.getenv("AZ_STORAGE_EP"),
  container_name = Sys.getenv("AZ_CONTAINER_INPUTS")
)
}
\arguments{
\item{ep_uri}{Character. Endpoint URI.}

\item{container_name}{Character. The name of the blob/storage container that
hosts files you want to read.}

\item{tenant}{Character. Tenant ID.}

\item{app_id}{Character. App ID.}
}
\value{
A blob_container/storage_container list.
}
\description{
Get Azure Container
}
