% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_age_sex_pyramid}
\alias{plot_age_sex_pyramid}
\title{Plot Age-Sex Pyramid}
\usage{
plot_age_sex_pyramid(age_sex_data)
}
\arguments{
\item{age_sex_data}{A data.frame. Read from Azure and processed with
\link{prepare_age_sex_data}. Counts for each strategy split by provider, year,
age group and sex.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Age-Sex Pyramid
}
