% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_rates.R
\name{plot_rates_box}
\alias{plot_rates_box}
\title{Plot Rates Boxplot with Peers}
\usage{
plot_rates_box(trend_data, plot_range)
}
\arguments{
\item{trend_data}{A data.frame.}

\item{plot_range}{Character. A vector of length two giving the y-axis limits.}
}
\value{
A ggplot2 object.
}
\description{
Plot Rates Boxplot with Peers
}
