% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_rates.R
\name{plot_rates_box}
\alias{plot_rates_box}
\title{Plot Rates Boxplot with Peers}
\usage{
plot_rates_box(trend_data, y_axis_limits)
}
\arguments{
\item{trend_data}{A data.frame.}

\item{y_axis_limits}{Numeric vector. Min and max values for the y axis.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Rates Boxplot with Peers
}
