% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_rates_box}
\alias{plot_rates_box}
\title{Plot Rates Boxplot with Peers}
\usage{
plot_rates_box(rates_box_data, y_axis_limits)
}
\arguments{
\item{rates_box_data}{A data.frame. Rates data read in from Azure and
processed with \link{generate_rates_baseline_data} to filter for provider,
strategy and year.}

\item{y_axis_limits}{Numeric vector. Min and max values for the y axis.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Rates Boxplot with Peers
}
