% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_rates_funnel}
\alias{plot_rates_funnel}
\title{Plot Rates Funnel with Peers}
\usage{
plot_rates_funnel(rates_funnel_data, y_axis_limits, x_axis_title)
}
\arguments{
\item{rates_funnel_data}{A data.frame. Rates data read in from Azure and
processed with \link{generate_rates_baseline_data} to filter for provider,
strategy and year, followed by \link{generate_rates_funnel_data} to generate
values for funnel structure.}

\item{y_axis_limits}{Numeric vector. Min and max values for the y axis.}

\item{x_axis_title}{Character. Title for the x-axis.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Rates Funnel with Peers
}
