% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_rates_trend}
\alias{plot_rates_trend}
\title{Plot Rates Trend Over Time}
\usage{
plot_rates_trend(
  rates_trend_data,
  selected_year,
  y_axis_limits,
  x_axis_title = "Financial year",
  y_axis_title,
  y_labels
)
}
\arguments{
\item{rates_trend_data}{A data.frame. Rates data read in from Azure, filtered
for a given provider and strategy, and arranged by year.}

\item{selected_year}{Integer. Selected year in the form \code{202324}.}

\item{y_axis_limits}{Numeric vector. Min and max values for the y axis.}

\item{x_axis_title}{Character. Title for the x-axis.}

\item{y_axis_title}{Character. Title for the y-axis.}

\item{y_labels}{A function. Function to format y-axis labels.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Rates Trend Over Time
}
