% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_rates_trend}
\alias{plot_rates_trend}
\title{Plot Rates Trend Over Time}
\usage{
plot_rates_trend(
  rates_trend_data,
  baseline_year = 202324,
  x_axis_title = "Financial year",
  y_axis_title = "Rate",
  y_axis_limits
)
}
\arguments{
\item{rates_trend_data}{A data.frame. Rates data read in from Azure, filtered
for a given provider and strategy, and arranged by year.}

\item{baseline_year}{Numeric. In the form \code{202324}.}

\item{x_axis_title}{Character. Title for the x-axis.}

\item{y_axis_title}{Character. Title for the y-axis.}

\item{y_axis_limits}{Numeric vector. Min and max values for the y axis.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Rates Trend Over Time
}
