% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_rates_trend}
\alias{plot_rates_trend}
\title{Plot Rates Trend Over Time}
\usage{
plot_rates_trend(
  rates_df,
  baseline_year = 202324,
  x_axis_title = "Financial year",
  y_axis_title = "Rate",
  y_axis_limits
)
}
\arguments{
\item{rates_df}{A data.frame. Must contain columns given by \code{fyear_col} and
\code{rate_col}. Pre-filtered for a given provider and strategy. One row per
financial year.}

\item{baseline_year}{Numeric. In the form \code{202324}.}

\item{x_axis_title}{Character.}

\item{y_axis_title}{Character.}

\item{y_axis_limits}{Numeric vector. Min and max values for the y axis.}
}
\value{
A 'ggplot2' object.
}
\description{
Plot Rates Trend Over Time
}
