% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_rates.R
\name{plot_rates_trend}
\alias{plot_rates_trend}
\title{Plot Rates Trend Over Time}
\usage{
plot_rates_trend(rates_df, fyear_col = "fyear", rate_col = "rate")
}
\arguments{
\item{rates_df}{A data.frame. Must contain columns given by \code{fyear_col} and
\code{rate_col}. Pre-filtered for a given provider and strategy. One row per
financial year.}

\item{fyear_col}{Character. Name of the column in \code{rates_df} containing the
financial year in the form \code{"2023/24"}.}

\item{rate_col}{Character. Name of the column in \code{rates_df} containing the
rate value (the type of which is dependent on the strategy).}
}
\value{
A ggplot2 object.
}
\description{
Plot Rates Trend Over Time
}
