% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{prepare_age_sex_data}
\alias{prepare_age_sex_data}
\title{Prepare Age-Sex Data}
\usage{
prepare_age_sex_data(age_sex_data)
}
\arguments{
\item{age_sex_data}{A data.frame. Read from Azure. Counts for each strategy
split by provider, year, age group and sex.}
}
\value{
A data.frame.
}
\description{
Prepare Age-Sex Data
}
