% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpma-explorer-app.R
\docType{package}
\name{tpma.explorer-package}
\alias{tpma.explorer}
\alias{tpma.explorer-package}
\title{tpma.explorer: Explore TPMA Data}
\description{
A Shiny-app-in-a-package to explore data related to Types of Potentially Mitigatable Activity (TPMAs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/The-Strategy-Unit/tpma-explorer}
  \item Report bugs at \url{https://github.com/The-Strategy-Unit/tpma-explorer/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{matt.dray@nhs.net}

}
\keyword{internal}
